chore: move initEventListener
